requirements/requirements_lock_3.10.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    bazel run //requirements:requirements_3.10.update
#
absl-py==2.3.1
    # via -r requirements/requirements.txt
async-timeout==5.0.1
    # via redis
attrs==25.3.0
    # via
    #   jsonschema
    #   osv
    #   referencing
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via requests
cffi==1.17.1
    # via pygit2
charset-normalizer==3.4.2
    # via requests
google-api-core[grpc]==2.25.1
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-datastore
    #   google-cloud-logging
    #   google-cloud-ndb
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-datastore
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-core==2.4.3
    # via
    #   google-cloud-datastore
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-datastore==2.21.0
    # via google-cloud-ndb
google-cloud-logging==3.12.1
    # via osv
google-cloud-ndb==2.3.4
    # via osv
google-cloud-storage==3.2.0
    # via -r requirements/requirements.txt
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-logging
grpcio==1.74.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   osv
grpcio-status==1.74.0
    # via google-api-core
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via opentelemetry-api
jinja2==3.1.6
    # via -r requirements/requirements.txt
jsonpath-ng==1.7.0
    # via -r requirements/requirements.txt
jsonschema==4.25.0
    # via osv
jsonschema-specifications==2025.4.1
    # via jsonschema
markupsafe==3.0.2
    # via jinja2
mmh3==5.2.0
    # via -r requirements/requirements.txt
opentelemetry-api==1.36.0
    # via google-cloud-logging
osv==0.0.21 ; python_version >= "3.10"
    # via -r requirements/requirements.txt
packaging==21.3
    # via osv
ply==3.11
    # via jsonpath-ng
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-datastore
    #   google-cloud-logging
protobuf==6.31.1 ; python_version >= "3.10"
    # via
    #   -r requirements/requirements.txt
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-datastore
    #   google-cloud-logging
    #   google-cloud-ndb
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pygit2==1.18.1
    # via osv
pymemcache==4.0.0
    # via google-cloud-ndb
pyparsing==3.2.3
    # via packaging
python-dateutil==2.9.0.post0
    # via -r requirements/requirements.txt
pytz==2025.2
    # via google-cloud-ndb
pyyaml==6.0.2
    # via osv
redis==6.4.0
    # via google-cloud-ndb
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   -r requirements/requirements.txt
    #   google-api-core
    #   google-cloud-storage
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
semver==3.0.4
    # via osv
six==1.17.0
    # via python-dateutil
tenacity==9.1.2
    # via -r requirements/requirements.txt
typing-extensions==4.14.1
    # via
    #   -r requirements/requirements.txt
    #   opentelemetry-api
    #   referencing
unidiff==0.7.5
    # via -r requirements/requirements.txt
urllib3==2.5.0
    # via requests
zipp==3.23.0
    # via importlib-metadata
